Created DAO Query utility. Worked on DAOImpl. Set up TableViews for Customer and Appointment tables. Refining DateTime Conversions and Formatting.

+++ src/application/model/User.java
@@ -7,7 +7,7 @@ public class User extends Record {
        ======================*/
     private String userName;
 
-    public User(long userID, String userName) {
+    public User(int userID, String userName) {
         super(userID);
         setUserName(userName);
     }
